akasa_air_task_v4/requirements.txt
pandas==2.2.2
sqlalchemy==2.0.32
pymysql==1.1.1
python-dotenv==1.0.1
lxml==5.3.0
tzdata==2024.1
pytest==8.2.0
